docs: Add hardening report for astraweave-render
- Documented pre-mutation-testing hardening efforts for `astraweave-render`.
- Included metrics: test count increased from 671 to 682, Clippy errors resolved from 10 to 0, and ensured clean compilation.
- Detailed critical correctness fixes, render path unification, and cleanup efforts.
- Highlighted remaining gaps and readiness for mutation testing.

+++ astraweave-render/src/nanite_render.rs
@@ -33,6 +33,7 @@ struct CameraUniform {
 }
 
 impl NaniteRenderContext {
+    #[allow(clippy::too_many_arguments)]
     pub fn new(
         device: &wgpu::Device,
         width: u32,
@@ -242,7 +243,7 @@ impl NaniteRenderContext {
         for &meshlet_id in &visible_meshlets {
             let meshlet = &meshlets[meshlet_id as usize];
             let vertex_count = meshlet.triangle_count * 3;
-            render_pass.draw(0..vertex_count as u32, meshlet_id..meshlet_id + 1);
+            render_pass.draw(0..vertex_count, meshlet_id..meshlet_id + 1);
         }
     }
